Moved the code from SymmRefiner and modifier their references in the Test.
